+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to `laravel-ddd` will be documented in this file.
 - The default domain model stub `model.php.stub` has changed. If stubs were published prior to this release, you may have to delete and re-publish; unless the published `model.php.stub` has been entirely customized with independent logic for your respective application.
 
 ### Changed
-- Implement proper support for custom base models when when using `ddd:model`:
+- Implement proper support for custom base models when using `ddd:model`:
     - If the configured `ddd.base_model` exists (evaluated using `class_exists`), base model generation is skipped.
     - If `ddd.base_model` does not exist and falls under a domain namespace, base model will be generated.
     - Falling under a domain namespace means `Domain\**\Models\**`.